Module traverse::utils

Structs

Counter
An infinite iterator starting at start and advancing by step with each iteration

Iterate
An iterator that repeatedly applies a given function, starting from a given seed value.

Range
An iterator over the range [start, stop)

RangeInclusive
An iterator over the range [start, stop]

RangeStep
An iterator over the range [start, stop) by step. It handles overflow by stopping.

RangeStepInclusive
An iterator over the range [start, stop] by step. It handles overflow by stopping.

Repeat
An iterator that repeats an element endlessly

Functions

count
Creates a new counter with the specified start/step

iterate
Create a new iterator that produces an infinite sequence of repeated applications of the given function f.

range
Returns an iterator over the given range [start, stop) (that is, starting at start (inclusive), and ending at stop (exclusive)).

range_inclusive
Return an iterator over the range [start, stop]

range_step
Return an iterator over the range [start, stop) by step. It handles overflow by stopping.

range_step_inclusive
Return an iterator over the range [start, stop] by step. It handles overflow by stopping.

repeat
Create a new iterator that endlessly repeats the element elt.