Function traverse::utils::iterate

pub fn iterate<T, F>(seed: T, f: F) -> Iterate<T, F> where T: Clone, F: FnMut(T) -> T

Create a new iterator that produces an infinite sequence of repeated applications of the given function f.